pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "taxembed"
version = "0.2.0"
description = "Hierarchical Poincaré embeddings for NCBI biological taxonomy"
readme = "README.md"
license = { text = "MIT" }
authors = [
    { name = "Joha Coludar", email = "jcoludar@gmail.com" }
]
requires-python = ">=3.11"
dependencies = [
    "torch>=2.0.0",
    "numpy>=1.21.0,<2.0",
    "pandas>=1.3.0",
    "scikit-learn>=1.0.0",
    "tqdm>=4.60.0",
    "matplotlib>=3.5.0",
    "umap-learn>=0.5.0",
]

[project.scripts]
taxembed-download = "taxembed.cli.download:main"
taxembed-prepare = "taxembed.cli.prepare:main"
taxembed-train = "taxembed.cli.train:main"
taxembed-visualize = "taxembed.cli.visualize:main"
taxembed-check = "taxembed.cli.check:main"

[project.optional-dependencies]
dev = [
    "ruff>=0.6.0",
    "pytest>=8.0.0",
    "pytest-cov>=4.1.0",
    "mypy>=1.0.0",
]

[project.urls]
Repository = "https://github.com/jcoludar/taxembed"
Documentation = "https://github.com/jcoludar/taxembed/blob/main/README.md"
Issues = "https://github.com/jcoludar/taxembed/issues"

[dependency-groups]
dev = [
    "ruff>=0.6.0",
    "pytest>=8.0.0",
    "pytest-cov>=4.1.0",
    "mypy>=1.0.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src/taxembed"]

[tool.ruff]
line-length = 100
target-version = "py311"
exclude = [
    ".git",
    ".venv",
    "venv",
    "venv311",
    "build",
    "dist",
    "__pycache__",
    ".pytest_cache",
    ".ruff_cache",
    "hype",  # Original Facebook code
]

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # pyflakes
    "I",      # isort
    "C",      # flake8-comprehensions
    "B",      # flake8-bugbear
    "UP",     # pyupgrade
]
ignore = [
    "E501",   # line too long (handled by formatter)
    "W503",   # line break before binary operator
    "B008",   # function calls in argument defaults
]

[tool.ruff.lint.isort]
known-first-party = ["taxembed"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "--cov=src/taxembed --cov-report=term-missing"

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false  # Gradual typing
